Android: fix the app build, and close the medium ATC and audio gaps
Two compile errors first, both from the previous commit and both only
findable in CI: speakChatter wants a Double and chatterSpeechLevel is a
Float, and distinctUntilChanged() applied straight to a StateFlow is a
deprecation *error* — a StateFlow already conflates equal values. Every
other call site in the app has a map, filter or combine ahead of it and
is fine. AGENTS.md records both, and its list of CI-only files was too
short: audio/, service/, map/, billing/, AppGraph and the Application
class are all outside :uicheck too.

Center frequencies. buildContext passed a constant 133.4, so only the
sector-to-sector crossing call ever used a real frequency: the Center
tune button and the Departure→Center hand-off named 133.4 whatever
sector the aircraft was in. It now reads the applied sector's own
frequency. The other defaults were wrong against iOS as well and are
spoken aloud in hand-offs — ground 121.9→121.8, departure
124.35→124.3, Center fallback 133.4→132.45.

The facility pill said a bare "Center" while centerSectorName sat
maintained, persisted, and already rendered in Settings. It now reads
"Fort Worth Center", falling back to the plain title before the sector
map resolves one.

The Tune Frequency grid was built from ATCFacility.entries, which
includes Ramp — so while tuned to Ramp the pilot got two Ramp buttons,
one from the grid and one AtcScreen adds itself. AtcFlowOrder.tunableFacilities
is the ported list, in the gate-to-gate order, deliberately without Ramp.

"Refresh airport data" omitted forceRefresh, so for any airport cached
inside 75 days it made no request at all — and it targeted the session
controller rather than the coordinator the taxi is routing on. Both now.
"Auto crossing calls" and "Auto-recalculate route" were read by nothing:
the coordinator kept its construction defaults whatever the pilot chose.

Audio. Tuning away now abandons the chatter exchange on the old
frequency — the loop reads the facility once per exchange, so a Tower
call and its read-back used to play out in full after the pilot had
switched to Ground. The Voice volume slider governed only the radio-effect
path: both clean-path speak() calls passed a null params Bundle, so with
"Transmission static" off the slider was inert and every call came out at
full volume. They now set KEY_PARAM_VOLUME.

Push-to-talk. The app synthesized the pilot's own read-back back at them
— they had just said it into the microphone. handleSpokenPilotText now
brackets its dispatch with pilotInputViaVoice and every pilot post goes
through shouldSpeakPilot. And "wilco" produced the full read-back,
reciting an instruction the pilot had deliberately not recited; wilco()
posts the acknowledgement, opens the read-back gate and tunes on a
hand-off. A runway-crossing clearance still takes the read-back path,
because on Android the words are what authorizes the crossing.

The golden transcript moves with the frequencies, and nothing else in it
changes.

73 of the 100 audited gaps are now closed and three partly closed; 11
medium and 13 low remain. :core is 1020 tests / 85 classes / 0 failures.

+++ Android/app/src/main/kotlin/com/h3consultingpartners/ifatccompanion/audio/AndroidChatterRadio.kt
@@ -89,7 +89,7 @@ class AndroidChatterRadio(
             // about 6 dB under the iOS level — thin beneath its own static, which sits at
             // the iOS level — and, worse, left it talking straight through a controller
             // clearance even with the duck applied, because the duck only reached the bed.
-            speech.speakChatter(line.spokenText, voiceId, engine.chatterSpeechLevel)
+            speech.speakChatter(line.spokenText, voiceId, engine.chatterSpeechLevel.toDouble())
         }
 
     /**
